# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Prospix API
  description: |
    API multi-tenant da plataforma de prospecção inteligente Prospix.
    Primeiro tenant: Giovane Carrara (MetLife · SJRP).

    ## Autenticação
    Todas as rotas (exceto `/auth/*` e `/webhooks/*`) exigem JWT Bearer + header `X-Tenant-Id`.
    O `X-Tenant-Id` é validado contra o JWT — mismatch retorna 403.

    ## Multi-tenancy
    Todo dado é isolado por tenant via RLS no PostgreSQL. Um tenant nunca vê dados de outro.

    ## Rate limits
    - `/auth/*`: 10 req/min por IP
    - `/tenant/*` GET: 600 req/min por tenant
    - `/tenant/*` POST/PATCH: 60 req/min por tenant
  version: 1.0.0
  contact:
    name: Guilds
    email: gustavo.macedo@guilds.com.br
servers:
  - url: https://api.prospix.com.br/v1
    description: Produção
  - url: https://api-staging.prospix.com.br/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local dev
...
security:
  - bearerAuth: []
    tenantHeader: []
paths:
...
  /auth/magic-link:
    post:
      tags: [Auth]
      summary: Solicita magic link
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                whatsapp: { type: string, example: "+5517998764422" }
                email: { type: string, format: email }
              oneOf:
                - required: [whatsapp]
                - required: [email]
      responses:
        '200':
          description: Link enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      sent_to: { type: string, enum: [whatsapp, email] }
                      expires_in: { type: integer, example: 600 }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/callback:
    get:
      tags: [Auth]
      summary: Callback do magic link (troca token por JWT)
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      access_token: { type: string }
                      refresh_token: { type: string }
                      user: { $ref: '#/components/schemas/User' }
                      tenant: { $ref: '#/components/schemas/Tenant' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renova access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      access_token: { type: string }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /auth/invitations/verify:
    post:
      tags: [Auth]
      summary: Verifica se código de convite é válido (sem consumir)
      x-prospix-response-shape: data-object
      x-prospix-error-shape: flat-error
      security: []
      description: |
        Rate limit 5 req/min por IP (anti brute-force).
        Em erro retorna mensagem genérica · `code` específico só visível no log interno.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: '^PRSPX-[A-Z0-9]{4}-[A-Z0-9]{4}$'
                  example: 'PRSPX-A8K2-9F4M'
      responses:
        '200':
          description: Código válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      tenant_name: { type: string, example: 'Giovane Carrara · MetLife · SJRP' }
                      role: { type: string, enum: [OWNER, ASSISTANT] }
        '404':
          description: INVITATION_INVALID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '410':
          description: INVITATION_EXPIRED ou INVITATION_ALREADY_USED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/invitations/redeem:
    post:
      tags: [Auth]
      summary: Consome código de convite + cria owner + envia magic link
      x-prospix-response-shape: data-object
      x-prospix-error-shape: flat-error
      security: []
      description: |
        Rate limit 3 req/min por IP.
        Marca invitation como usado e dispara magic link via WhatsApp do owner.
        Operação atômica: ou cria tudo, ou nada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, user, accept_terms]
              properties:
                code:
                  type: string
                  pattern: '^PRSPX-[A-Z0-9]{4}-[A-Z0-9]{4}$'
                user:
                  type: object
                  required: [name, email, whatsapp]
                  properties:
                    name: { type: string, maxLength: 255 }
                    email: { type: string, format: email }
                    whatsapp: { type: string, example: '+5517998764422' }
                    susep: { type: string, nullable: true }
                    city: { type: string, nullable: true }
                accept_terms:
                  type: boolean
                  description: Aceite LGPD + Termos de uso (obrigatório true)
      responses:
        '201':
          description: Cadastrado · magic link enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user_id: { type: string, format: uuid }
                      tenant_id: { type: string, format: uuid }
                      magic_link_sent: { type: boolean, example: true }
                      sent_to: { type: string, example: 'whatsapp' }
        '410':
          description: Código expirado ou já usado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/admin-login:
    post:
      tags: [Auth]
      summary: Login administrativo Guilds
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 1 }
      responses:
        '200':
          description: Sessão administrativa criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  user: { type: object }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { description: Credenciais administrativas inválidas }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
  responses:
    Unauthenticated:
      description: Token ausente ou inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Payload inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Limite de requisições excedido
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [UNAUTHENTICATED, UNAUTHORIZED, RESOURCE_NOT_FOUND, VALIDATION_ERROR, RATE_LIMITED, TENANT_QUOTA_EXCEEDED, EXTERNAL_SERVICE_DOWN, INTERNAL_ERROR]
            message: { type: string }
            details: { type: object }
            trace_id: { type: string }
        meta:
          type: object
          properties:
            request_id: { type: string }
            timestamp: { type: string, format: date-time }
...
    Tenant:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        plan: { type: string }
        brand_primary_color: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [owner, assistant, guilds_admin] }
        name: { type: string }
        email: { type: string, format: email }
        whatsapp: { type: string }
        susep: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - UNAUTHORIZED
                - RESOURCE_NOT_FOUND
                - VALIDATION_ERROR
                - RATE_LIMITED
                - TENANT_QUOTA_EXCEEDED
                - INVITATION_INVALID
                - INVITATION_EXPIRED
                - INVITATION_ALREADY_USED
                - EXTERNAL_SERVICE_DOWN
                - INTERNAL_ERROR
            message: { type: string }
            details: { type: object, additionalProperties: true }
            trace_id: { type: string, nullable: true }
        meta:
          type: object
          properties:
            request_id: { type: string }
            timestamp: { type: string, format: date-time }